Added new library componet project to re-organized the structure of the project
