+++ CustomerAPI.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29806.167
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomerAPI", "CustomerAPI.csproj", "{A45B4679-4C25-4C39-975B-BBE9D61C7B46}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomerAPI", "CustomerAPI.csproj", "{A45B4679-4C25-4C39-975B-BBE9D61C7B46}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Customer.Components", "..\Customer.Components\Customer.Components.csproj", "{45938FA9-BF7C-488D-A58A-41D99C24674D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{A45B4679-4C25-4C39-975B-BBE9D61C7B46}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A45B4679-4C25-4C39-975B-BBE9D61C7B46}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A45B4679-4C25-4C39-975B-BBE9D61C7B46}.Release|Any CPU.Build.0 = Release|Any CPU
+		{45938FA9-BF7C-488D-A58A-41D99C24674D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{45938FA9-BF7C-488D-A58A-41D99C24674D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{45938FA9-BF7C-488D-A58A-41D99C24674D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{45938FA9-BF7C-488D-A58A-41D99C24674D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
