--- Dtos/PagingDto.cs
@@ -1,51 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace CustomerAPI.Dtos
-{
-    /// <summary>
-    /// The Paging DTO
-    /// </summary>
-    /// <typeparam name="T">Generic Model</typeparam>
-    public class PagingDto<T> where T : class
-    {
-        /// <summary>
-        /// Gets or sets the collection.
-        /// </summary>
-        /// <value>
-        /// The collection.
-        /// </value>
-        public IList<T> Collection { get; set; }
-
-        /// <summary>
-        /// Gets or sets the index of the page.
-        /// </summary>
-        /// <value>
-        /// The index of the page.
-        /// </value>
-        public int PageIndex { get; set; }
-
-        /// <summary>
-        /// Gets or sets the total pages.
-        /// </summary>
-        /// <value>
-        /// The total pages.
-        /// </value>
-        public int TotalPages { get; set; }
-
-        /// <summary>
-        /// Gets or sets the total records.
-        /// </summary>
-        /// <value>
-        /// The total records.
-        /// </value>
-        public long TotalRecords { get; set; }
-
-        /// <summary>
-        /// Gets or sets the size of the page.
-        /// </summary>
-        /// <value>
-        /// The size of the page.
-        /// </value>
-        public int PageSize { get; set; }
-    }
-}
