--- Dtos/ProblemDetailsDto.cs
@@ -1,66 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace CustomerAPI.Dtos
-{
-    /// <summary>
-    /// The problem details dto.
-    /// </summary>
-    public class ProblemDetailsDto
-    {
-        /// <summary>
-        /// Gets or sets the error code.
-        /// </summary>
-        /// <value>
-        /// The error code.
-        /// </value>
-        public int? ErrorCode { get; set; }
-
-        /// <summary>
-        /// Gets or sets the title.
-        /// </summary>
-        /// <value>
-        /// The title.
-        /// </value>
-        public string Title { get; set; }
-
-        /// <summary>
-        /// Gets or sets the detail.
-        /// </summary>
-        /// <value>
-        /// The detail.
-        /// </value>
-        public string Detail { get; set; }
-
-        /// <summary>
-        /// Gets or sets the type.
-        /// </summary>
-        /// <value>
-        /// The type.
-        /// </value>
-        public string Type { get; set; }
-
-        /// <summary>
-        /// Gets or sets the instance.
-        /// </summary>
-        /// <value>
-        /// The instance.
-        /// </value>
-        public string Instance { get; set; }
-
-        /// <summary>
-        /// Gets or sets the status.
-        /// </summary>
-        /// <value>
-        /// The status.
-        /// </value>
-        public int? Status { get; set; }
-
-        /// <summary>
-        /// Gets or sets the extensions.
-        /// </summary>
-        /// <value>
-        /// The extensions.
-        /// </value>
-        public IDictionary<string, object> Extensions { get; set; }
-    }
-}
