--- Dtos/Responses/CustomerDto.cs
@@ -1,50 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace CustomerAPI.Dtos.Responses
-{
-    /// <summary>
-    /// The Customer DTO
-    /// </summary>
-    public class CustomerDto
-    {
-        /// <summary>
-        /// Gets or sets the identifier.
-        /// </summary>
-        /// <value>
-        /// The identifier.
-        /// </value>
-        public string Id { get; set; }
-
-        /// <summary>
-        /// Gets or sets the full name.
-        /// </summary>
-        /// <value>
-        /// The full name.
-        /// </value>
-        public string FullName { get; set; }
-
-        /// <summary>
-        /// Gets or sets the date of birth.
-        /// </summary>
-        /// <value>
-        /// The date of birth.
-        /// </value>
-        public string DateOfBirth { get; set; }
-
-        /// <summary>
-        /// Gets or sets the age.
-        /// </summary>
-        /// <value>
-        /// The age.
-        /// </value>
-        public int Age { get; set; }
-
-        /// <summary>
-        /// Gets or sets the address.
-        /// </summary>
-        /// <value>
-        /// The address.
-        /// </value>
-        public ICollection<AddressDto> Address { get; set; }
-    }
-}
