--- Requests/AddressRequest.cs
@@ -1,12 +1,0 @@
-﻿using CustomerAPI.Dtos;
-
-namespace CustomerAPI.Requests
-{
-    /// <summary>
-    /// Create Address Request
-    /// </summary>
-    /// <seealso cref="CustomerAPI.Dtos.AddressDto"/>
-    public class AddressRequest : AddressDto
-    {
-    }
-}
